pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "universal-debug-agent"
version = "0.1.0"
description = "Universal debug/investigation agent based on OpenAI Agents SDK"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "openai-agents>=0.0.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
debug-agent = "universal_debug_agent.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/universal_debug_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
